refactor test6.cpp
